+++ src/result/ChargeResult.java
@@ -2,6 +2,6 @@ package result;
 
 public class ChargeResult extends PhaseResult{
     public ChargeResult(String phaseName) {
-        super("Charge");
+        super(PhaseType.CHARGE);
     }
 }