+++ src/result/PhaseType.java
@@ -1,0 +1,8 @@
+package result;
+
+public enum PhaseType {
+    BATTLE,
+    SHOOTING,
+    FIGHT,
+    CHARGE
+}